Update CUB, and fix bugs
 All thrust unit tests & examples passes through cuda-memcheck successfully

[git-p4: depot-paths = "//sw/gpgpu/thrust/": change = 21437187]

+++ thrust/system/cuda/detail/cub/block/specializations/block_reduce_warp_reductions.cuh
@@ -94,9 +94,9 @@ struct BlockReduceWarpReductions
 
     // Thread fields
     _TempStorage &temp_storage;
-    int linear_tid;
-    int warp_id;
-    int lane_id;
+    unsigned int linear_tid;
+    unsigned int warp_id;
+    unsigned int lane_id;
 
 
     /// Constructor
@@ -127,9 +127,9 @@ struct BlockReduceWarpReductions
 
     template <bool FULL_TILE, typename ReductionOp>
     __device__ __forceinline__ T ApplyWarpAggregates(
-        ReductionOp         /*reduction_op*/,       ///< [in] Binary scan operator
+        ReductionOp         /*reduction_op*/,   ///< [in] Binary scan operator
         T                   warp_aggregate,     ///< [in] <b>[<em>lane</em><sub>0</sub> only]</b> Warp-wide aggregate reduction of input items
-        int                 /*num_valid*/,          ///< [in] Number of valid elements (may be less than BLOCK_THREADS)
+        int                 /*num_valid*/,      ///< [in] Number of valid elements (may be less than BLOCK_THREADS)
         Int2Type<WARPS>     /*successor_warp*/)
     {
         return warp_aggregate;
@@ -197,10 +197,10 @@ struct BlockReduceWarpReductions
         int                 num_valid,          ///< [in] Number of valid elements (may be less than BLOCK_THREADS)
         ReductionOp         reduction_op)       ///< [in] Binary reduction operator
     {
-        int    warp_offset = warp_id * LOGICAL_WARP_SIZE;
-        int    warp_num_valid = (FULL_TILE && EVEN_WARP_MULTIPLE) ?
+        unsigned int    warp_offset = warp_id * LOGICAL_WARP_SIZE;
+        unsigned int    warp_num_valid = (FULL_TILE && EVEN_WARP_MULTIPLE) ?
                             LOGICAL_WARP_SIZE :
-                            (warp_offset < num_valid) ?
+                            (warp_offset < static_cast<unsigned int>(num_valid)) ?
                                 num_valid - warp_offset :
                                 0;
